requirements.txt
streamlit
opencv-python
numpy
matplotlib
Pillow
torch
torchvision
kornia